removed last aux_func
